% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCLUST.R
\name{HCLUST}
\alias{HCLUST}
\title{Hierarchical Clustering Method}
\usage{
HCLUST(data.matrix)
}
\arguments{
\item{data.matrix}{The data need to be analyzed}
}
\value{
The cluster index matrix
}
\description{
Hierarchical Clustering Method
}
\examples{
HCLUST(data = as.matrix(iris[,1:4]))
}
