% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{My Function}
\usage{
hello(name)
}
\arguments{
\item{name}{The name of the person to say hi to}
}
\value{
The ouput from \code{\link{print}}
}
\description{
My Function
}
\examples{
hello("CS599")
\dontrun{
hello("J")
}
}
