% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMEANS.R
\name{KMEANS}
\alias{KMEANS}
\title{KMEANS function}
\usage{
KMEANS(data = NA, k = NA)
}
\arguments{
\item{data}{The name of the data need to be anlyzied}

\item{k}{The number of clusters}
}
\value{
the clusters
}
\description{
KMEANS function
}
\examples{
KMEANS(data = as.matrix(iris[,1:4]), k =3)
}
